// types/types.ts
export type Advertisement = {
    id?: string | number
    advertisementId?: string | number
    name?: string
    title?: string
    description?: string
    agencyName?: string
    agency?: string
    agencyAccountName?: string
    price?: string | number
    amount?: string | number
    listingPrice?: string | number
    priceEuro?: string | number
}